--- MakFood.Kitchen/MakFood.Kitchen.Domain/Base/BaseEntity.cs
@@ -1,13 +1,0 @@
-﻿namespace MakFood.Kitchen.Domain.Base
-{
-    public abstract class BaseEntity<TId>
-    {
-        protected BaseEntity()
-        {
-            CreationDateTime = DateTime.Now;
-        }
-
-        public TId Id { get; protected set; }
-        public DateTime CreationDateTime { get; private set; }
-    }
-}
