--- MakFood.Kitchen/MakFood.Kitchen.Domain/DiscountAggrigate/DiscountPolicyAggrigate/AllPermittedPolicy.cs
@@ -1,14 +1,0 @@
-﻿using MakFood.Kitchen.Domain.DiscountAggrigate.Enum;
-
-namespace MakFood.Kitchen.Domain.DiscountAggrigate.DiscountPolicyAggrigate
-{
-    public class AllPermittedPolicy :  DiscountPolicy 
-    {
-        public AllPermittedPolicy() : base(DiscountPolicyType.AllPermitted)
-        {
-        }
-
-        public override bool IsPermitted(Guid CamtomerId)=>true;
-    }
-}
-
