--- MakFood.Kitchen/MakFood.Kitchen.Domain/DiscountAggrigate/DiscountPolicyAggrigate/SpecifiedPermisionPolicy.cs
@@ -1,43 +1,0 @@
-﻿using MakFood.Kitchen.Domain.DiscountAggrigate.Enum;
-using MakFood.Kitchen.Infrastructure.Substructure.Extensions;
-
-namespace MakFood.Kitchen.Domain.DiscountAggrigate.DiscountPolicyAggrigate
-{
-    public class SpecifiedPermisionPolicy : DiscountPolicy
-    {
-        private List<Guid> _customers = new List<Guid>();
-        public SpecifiedPermisionPolicy(IEnumerable<Guid> customerIds) : base(DiscountPolicyType.SpecifiedPermision)
-        {
-
-            _customers = customerIds.ToList() ?? new List<Guid>();
-        }
-
-        public IEnumerable<Guid> Customers => _customers.AsReadOnly();
-        
-        #region Behavior
-        public override bool IsPermitted(Guid customerId)
-        {
-            CustomeridValidator(customerId);
-            return Customers.Contains(customerId);
-        }
-
-        public void AddCustomers(Guid customerId)
-        {
-            CustomeridValidator(customerId);
-            _customers.Add(customerId);
-        }
-        #endregion
-
-        #region Validations
-
-        private void CustomeridValidator(Guid customerId)
-        {
-            customerId.CheckNullOrEmpty("Customer Id");
-        }
-
-        #endregion
-
-
-
-    }
-}
