--- MakFood.Kitchen/MakFood.Kitchen.Domain/DiscountAggrigate/Enum/DiscountPolicyType.cs
@@ -1,8 +1,0 @@
-﻿namespace MakFood.Kitchen.Domain.DiscountAggrigate.Enum
-{
-    public enum DiscountPolicyType 
-    {
-        AllPermitted,
-        SpecifiedPermision
-    }
-}
