+++ MakFood.Kitchen/MakFood.Kitchen.Domain/Entities/DiscountAggrigate/Enum/DiscountPolicyType.cs
@@ -1,0 +1,8 @@
+﻿namespace MakFood.Kitchen.Domain.Entities.DiscountAggrigate.Enum
+{
+    public enum DiscountPolicyType 
+    {
+        AllPermitted,
+        SpecifiedPermision
+    }
+}
