+++ MakFood.Kitchen/MakFood.Kitchen.Domain/Entities/OrderAggrigate/OrderAggrigate/OrederState/MiseOnPlaceOrderState.cs
@@ -1,0 +1,10 @@
+﻿namespace MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.OrederState
+{
+    public sealed class MiseOnPlaceOrderState : OrderState
+    {
+        internal MiseOnPlaceOrderState() { }
+        public override OrderStatus Status => OrderStatus.MiseOnPlace;
+
+    }
+
+}
