+++ MakFood.Kitchen/MakFood.Kitchen.Domain/Entities/OrderAggrigate/OrderAggrigate/PaymentAggrigate/Enum/PaymentMathods.cs
@@ -1,0 +1,8 @@
+﻿namespace MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.PaymentAggrigate.Enum
+{
+    public enum PaymentMathods
+    {
+        Cash,
+        Wallet
+    }
+}
