+++ MakFood.Kitchen/MakFood.Kitchen.Domain/Entities/OrderAggrigate/OrderAggrigate/PaymentAggrigate/Enum/PaymentStatus.cs
@@ -1,0 +1,9 @@
+﻿namespace MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.PaymentAggrigate.Enum
+{
+    public enum PaymentStatus
+    {
+        Created,
+        Paid,
+        Cancelled
+    }
+}
