+++ MakFood.Kitchen/MakFood.Kitchen.Domain/Entities/OrderAggrigate/OrderAggrigate/PaymentAggrigate/SingelePayment.cs
@@ -1,0 +1,18 @@
+﻿using MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.PaymentAggrigate.Enum;
+using MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.PaymentAggrigate.PaymentBase;
+
+namespace MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.PaymentAggrigate
+{
+    public class SingelePayment : Payment
+    {
+        public SingelePayment(decimal totalAmount, decimal reminingAmount, PaymentMathods ownerPaymentMethod, decimal ownerAmount,
+            decimal ownerPaidAmount) : base(totalAmount, reminingAmount, ownerPaymentMethod, ownerAmount, ownerPaidAmount)
+        {
+            TotalAmount = totalAmount;
+            ReminingAmount = reminingAmount;
+            OwnerPaymentMethod = ownerPaymentMethod;
+            OwnerAmount = ownerAmount;
+            OwnerPaidAmount = ownerPaidAmount;
+        }
+    }
+}
