+++ MakFood.Kitchen/MakFood.Kitchen.Domain/Entities/OrderAggrigate/OrderAggrigate/PaymentAggrigate/State/CancelledPaymentState.cs
@@ -1,0 +1,9 @@
+﻿using MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.PaymentAggrigate.Enum;
+
+namespace MakFood.Kitchen.Domain.Entities.OrderAggrigate.OrderAggrigate.PaymentAggrigate.State
+{
+    public sealed class CancelledPaymentState : PaymentState
+    {
+        public override PaymentStatus Status => PaymentStatus.Cancelled;
+    }
+}
