--- MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/Enum/OrderStatus.cs
@@ -1,11 +1,0 @@
-﻿namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.Enum
-{
-    public enum OrderStatus
-    { 
-        InvoiceSent,
-        Paid,
-        MiseOnPlace,
-        Cancelled
-
-    }
-}
