--- MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/OrederState/OrderStatus.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.OrederState
-{
-    public enum OrderStatus
-    {
-        Created,
-        MiseOnPlace,
-        Cancelled
-    }
-}
