--- MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/PaymentAggrigate/Enum/PaymentMathods.cs
@@ -1,8 +1,0 @@
-﻿namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.PaymentAggrigate.Enum
-{
-    public enum PaymentMathods
-    {
-        Cash,
-        Wallet
-    }
-}
