--- MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/PaymentAggrigate/Enum/PaymentStatus.cs
@@ -1,9 +1,0 @@
-﻿namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.PaymentAggrigate.Enum
-{
-    public enum PaymentStatus
-    {
-        Created,
-        Paid,
-        Cancelled
-    }
-}
