--- MakFood.Kitchen/MakFood.Kitchen.Domain/OrderAggrigate/OrderAggrigate/PaymentAggrigate/State/CreatedPaymentState.cs
@@ -1,19 +1,0 @@
-﻿using MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.PaymentAggrigate.Enum;
-
-namespace MakFood.Kitchen.Domain.OrderAggrigate.OrderAggrigate.PaymentAggrigate.State
-{
-    public class CreatedPaymentState : PaymentState
-    {
-        public override PaymentStatus Status => PaymentStatus.Created;
-
-        public override PaidPaymentState Paid()
-        {
-            return new PaidPaymentState();
-        }
-
-        public override CancelledPaymentState Cancelled()
-        {
-            return new CancelledPaymentState();
-        }
-    }
-}
